Delete old backend
