+++ backend/columnar.js
@@ -696,17 +696,6 @@ function decodeContainerHeader(decoder, computeHash) {
   return header
 }
 
-/**
- * Returns the checksum of a change (bytes 4 to 7) as a 32-bit unsigned integer.
- */
-function getChangeChecksum(change) {
-  if (change[0] !== MAGIC_BYTES[0] || change[1] !== MAGIC_BYTES[1] ||
-      change[2] !== MAGIC_BYTES[2] || change[3] !== MAGIC_BYTES[3]) {
-    throw new RangeError('Data does not begin with magic bytes 85 6f 4a 83')
-  }
-  return ((change[4] << 24) | (change[5] << 16) | (change[6] << 8) | change[7]) >>> 0
-}
-
 function encodeChange(changeObj) {
   const { changes, actorIds } = parseAllOpIds([changeObj], true)
   const change = changes[0]
@@ -868,76 +857,6 @@ function sortOpIds(a, b) {
   return 0
 }
 
-function groupDocumentOps(changes) {
-  let byObjectId = {}, byReference = {}, objectType = {}
-  for (let change of changes) {
-    for (let i = 0; i < change.ops.length; i++) {
-      const op = change.ops[i], opId = `${op.id.counter}@${op.id.actorId}`
-      const objectId = (op.obj === '_root') ? '_root' : `${op.obj.counter}@${op.obj.actorId}`
-      if (op.action.startsWith('make')) {
-        objectType[opId] = op.action
-        if (op.action === 'makeList' || op.action === 'makeText') {
-          byReference[opId] = {'_head': []}
-        }
-      }
-
-      let key
-      if (objectId === '_root' || objectType[objectId] === 'makeMap' || objectType[objectId] === 'makeTable') {
-        key = op.key
-      } else if (objectType[objectId] === 'makeList' || objectType[objectId] === 'makeText') {
-        if (op.insert) {
-          key = opId
-          const ref = (op.elemId === '_head') ? '_head' : `${op.elemId.counter}@${op.elemId.actorId}`
-          byReference[objectId][ref].push(opId)
-          byReference[objectId][opId] = []
-        } else {
-          key = `${op.elemId.counter}@${op.elemId.actorId}`
-        }
-      } else {
-        throw new RangeError(`Unknown object type for object ${objectId}`)
-      }
-
-      if (!byObjectId[objectId]) byObjectId[objectId] = {}
-      if (!byObjectId[objectId][key]) byObjectId[objectId][key] = {}
-      byObjectId[objectId][key][opId] = op
-      op.succ = []
-
-      for (let pred of op.pred) {
-        const predId = `${pred.counter}@${pred.actorId}`
-        if (!byObjectId[objectId][key][predId]) {
-          throw new RangeError(`No predecessor operation ${predId}`)
-        }
-        byObjectId[objectId][key][predId].succ.push(op.id)
-      }
-    }
-  }
-
-  let ops = []
-  for (let objectId of Object.keys(byObjectId).sort(sortOpIds)) {
-    let keys = []
-    if (objectType[objectId] === 'makeList' || objectType[objectId] === 'makeText') {
-      let stack = ['_head']
-      while (stack.length > 0) {
-        const key = stack.pop()
-        if (key !== '_head') keys.push(key)
-        for (let opId of byReference[objectId][key].sort(sortOpIds)) stack.push(opId)
-      }
-    } else {
-      // FIXME JavaScript sorts based on UTF-16 encoding. We should change this to use the UTF-8
-      // encoding instead (the sort order will be different beyond the basic multilingual plane)
-      keys = Object.keys(byObjectId[objectId]).sort()
-    }
-
-    for (let key of keys) {
-      for (let opId of Object.keys(byObjectId[objectId][key]).sort(sortOpIds)) {
-        const op = byObjectId[objectId][key][opId]
-        if (op.action !== 'del') ops.push(op)
-      }
-    }
-  }
-  return ops
-}
-
 /**
  * Takes a set of operations `ops` loaded from an encoded document, and
  * reconstructs the changes that they originally came from.
@@ -1010,60 +929,6 @@ function groupChangeOps(changes, ops) {
       delete op.id
     }
   }
-}
-
-function encodeDocumentChanges(changes) {
-  const columns = { // see DOCUMENT_COLUMNS
-    actor     : new RLEEncoder('uint'),
-    seq       : new DeltaEncoder(),
-    maxOp     : new DeltaEncoder(),
-    time      : new DeltaEncoder(),
-    message   : new RLEEncoder('utf8'),
-    depsNum   : new RLEEncoder('uint'),
-    depsIndex : new DeltaEncoder(),
-    extraLen  : new RLEEncoder('uint'),
-    extraRaw  : new Encoder()
-  }
-  let indexByHash = {} // map from change hash to its index in the changes array
-  let heads = {} // change hashes that are not a dependency of any other change
-
-  for (let i = 0; i < changes.length; i++) {
-    const change = changes[i]
-    indexByHash[change.hash] = i
-    heads[change.hash] = true
-
-    columns.actor.appendValue(change.actorNum)
-    columns.seq.appendValue(change.seq)
-    columns.maxOp.appendValue(change.startOp + change.ops.length - 1)
-    columns.time.appendValue(change.time)
-    columns.message.appendValue(change.message)
-    columns.depsNum.appendValue(change.deps.length)
-
-    for (let dep of change.deps) {
-      if (typeof indexByHash[dep] !== 'number') {
-        throw new RangeError(`Unknown dependency hash: ${dep}`)
-      }
-      columns.depsIndex.appendValue(indexByHash[dep])
-      if (heads[dep]) delete heads[dep]
-    }
-
-    if (change.extraBytes) {
-      columns.extraLen.appendValue(change.extraBytes.byteLength << 4 | VALUE_TYPE.BYTES)
-      columns.extraRaw.appendRawBytes(change.extraBytes)
-    } else {
-      columns.extraLen.appendValue(VALUE_TYPE.BYTES) // zero-length byte array
-    }
-  }
-
-  let changesColumns = []
-  for (let {columnName, columnId} of DOCUMENT_COLUMNS) {
-    changesColumns.push({columnId, columnName, encoder: columns[columnName]})
-  }
-  changesColumns.sort((a, b) => a.columnId - b.columnId)
-
-  const sortedHeads = Object.keys(heads).sort()
-  const headsIndexes = sortedHeads.map(hash => indexByHash[hash])
-  return { changesColumns, heads: sortedHeads, headsIndexes }
 }
 
 function decodeDocumentChanges(changes, expectedHeads) {
@@ -1161,16 +1026,6 @@ function decodeDocumentHeader(buffer) {
   return { changesColumns, opsColumns, actorIds, heads, headsIndexes, extraBytes }
 }
 
-/**
- * Transforms a list of changes into a binary representation of the document state.
- */
-function encodeDocument(binaryChanges) {
-  const { changes, actorIds } = parseAllOpIds(decodeChanges(binaryChanges), false)
-  const { changesColumns, heads, headsIndexes } = encodeDocumentChanges(changes)
-  const opsColumns = encodeOps(groupDocumentOps(changes), true)
-  return encodeDocumentHeader({ changesColumns, opsColumns, actorIds, heads, headsIndexes })
-}
-
 function decodeDocument(buffer) {
   const { changesColumns, opsColumns, actorIds, heads } = decodeDocumentHeader(buffer)
   const changes = decodeColumns(changesColumns, actorIds, DOCUMENT_COLUMNS)
@@ -1200,229 +1055,9 @@ function inflateColumn(column) {
   }
 }
 
-/**
- * Takes all the operations for the same property (i.e. the same key in a map, or the same list
- * element) and mutates the object patch to reflect the current value(s) of that property. There
- * might be multiple values in the case of a conflict. `objects` is a map from objectId to the
- * patch for that object. `property` contains `objId`, `key`, a list of `ops`, and `index` (the
- * current list index if the object is a list). Returns true if one or more values are present,
- * or false if the property has been deleted.
- */
-function addPatchProperty(objects, property) {
-  let values = {}, counter = null
-  for (let op of property.ops) {
-    // Apply counters and their increments regardless of the number of successor operations
-    if (op.actionName === 'set' && op.value.datatype === 'counter') {
-      if (!counter) counter = {opId: op.opId, value: 0, succ: {}}
-      counter.value += op.value.value
-      for (let succId of op.succ) counter.succ[succId] = true
-    } else if (op.actionName === 'inc') {
-      if (!counter) throw new RangeError(`inc operation ${op.opId} without a counter`)
-      counter.value += op.value.value
-      delete counter.succ[op.opId]
-      for (let succId of op.succ) counter.succ[succId] = true
-
-    } else if (op.succ.length === 0) { // Ignore any ops that have been overwritten
-      if (op.actionName.startsWith('make')) {
-        values[op.opId] = objects[op.opId]
-      } else if (op.actionName === 'set') {
-        values[op.opId] = {value: op.value.value, type: 'value'}
-        if (op.value.datatype) {
-          values[op.opId].datatype = op.value.datatype
-        }
-      } else if (op.actionName === 'link') {
-        // NB. This assumes that the ID of the child object is greater than the ID of the current
-        // object. This is true as long as link operations are only used to redo undone make*
-        // operations, but it will cease to be true once subtree moves are allowed.
-        if (!op.childId) throw new RangeError(`link operation ${op.opId} without a childId`)
-        values[op.opId] = objects[op.childId]
-      } else {
-        throw new RangeError(`Unexpected action type: ${op.actionName}`)
-      }
-    }
-  }
-
-  // If the counter had any successor operation that was not an increment, that means the counter
-  // must have been deleted, so we omit it from the patch.
-  if (counter && Object.keys(counter.succ).length === 0) {
-    values[counter.opId] = {type: 'value', value: counter.value, datatype: 'counter'}
-  }
-
-  if (Object.keys(values).length > 0) {
-    let obj = objects[property.objId]
-    if (obj.type === 'map' || obj.type === 'table') {
-      obj.props[property.key] = values
-    } else if (obj.type === 'list' || obj.type === 'text') {
-      makeListEdits(obj, values, property.key, property.index)
-    }
-    return true
-  } else {
-    return false
-  }
-}
-
-/**
- * When constructing a patch to instantiate a loaded document, this function adds the edits to
- * insert one list element. Usually there is one value, but in the case of a conflict there may be
- * several values. `elemId` is the ID of the list element, and `index` is the list index at which
- * the value(s) should be placed.
- */
-function makeListEdits(list, values, elemId, index) {
-  let firstValue = true
-  const opIds = Object.keys(values).sort((id1, id2) => compareParsedOpIds(parseOpId(id1), parseOpId(id2)))
-  for (const opId of opIds) {
-    if (firstValue) {
-      list.edits.push({action: 'insert', value: values[opId], elemId, opId, index})
-    } else {
-      list.edits.push({action: 'update', value: values[opId], opId, index})
-    }
-    firstValue = false
-  }
-}
-
-/**
- * Recursively walks the patch tree, calling appendEdit on every list edit in order to consense
- * consecutive sequences of insertions into multi-inserts.
- */
-function condenseEdits(diff) {
-  if (diff.type === 'list' || diff.type === 'text') {
-    diff.edits.forEach(e => condenseEdits(e.value))
-    let newEdits = diff.edits
-    diff.edits = []
-    for (const edit of newEdits) appendEdit(diff.edits, edit)
-  } else if (diff.type === 'map' || diff.type === 'table') {
-    for (const prop of Object.keys(diff.props)) {
-      for (const opId of Object.keys(diff.props[prop])) {
-        condenseEdits(diff.props[prop][opId])
-      }
-    }
-  }
-}
-
-/**
- * Appends a list edit operation (insert, update, remove) to an array of existing operations. If the
- * last existing operation can be extended (as a multi-op), we do that.
- */
-function appendEdit(existingEdits, nextEdit) {
-  if (existingEdits.length === 0) {
-    existingEdits.push(nextEdit)
-    return
-  }
-
-  let lastEdit = existingEdits[existingEdits.length - 1]
-  if (lastEdit.action === 'insert' && nextEdit.action === 'insert' &&
-      lastEdit.index === nextEdit.index - 1 &&
-      lastEdit.value.type === 'value' && nextEdit.value.type === 'value' &&
-      lastEdit.elemId === lastEdit.opId && nextEdit.elemId === nextEdit.opId &&
-      opIdDelta(lastEdit.elemId, nextEdit.elemId, 1)) {
-    lastEdit.action = 'multi-insert'
-    lastEdit.values = [lastEdit.value.value, nextEdit.value.value]
-    delete lastEdit.value
-    delete lastEdit.opId
-
-  } else if (lastEdit.action === 'multi-insert' && nextEdit.action === 'insert' &&
-             lastEdit.index + lastEdit.values.length === nextEdit.index &&
-             nextEdit.value.type === 'value' && nextEdit.elemId === nextEdit.opId &&
-             opIdDelta(lastEdit.elemId, nextEdit.elemId, lastEdit.values.length)) {
-    lastEdit.values.push(nextEdit.value.value)
-
-  } else if (lastEdit.action === 'remove' && nextEdit.action === 'remove' &&
-             lastEdit.index === nextEdit.index) {
-    lastEdit.count += nextEdit.count
-
-  } else {
-    existingEdits.push(nextEdit)
-  }
-}
-
-/**
- * Returns true if the two given operation IDs have the same actor ID, and the counter of `id2` is
- * exactly `delta` greater than the counter of `id1`.
- */
-function opIdDelta(id1, id2, delta = 1) {
-  const parsed1 = parseOpId(id1), parsed2 = parseOpId(id2)
-  return parsed1.actorId === parsed2.actorId && parsed1.counter + delta === parsed2.counter
-}
-
-/**
- * Parses the document (in compressed binary format) given as `documentBuffer`
- * and returns a patch that can be sent to the frontend to instantiate the
- * current state of that document.
- */
-function constructPatch(documentBuffer) {
-  const { opsColumns, actorIds } = decodeDocumentHeader(documentBuffer)
-  const col = makeDecoders(opsColumns, DOC_OPS_COLUMNS).reduce(
-    (acc, col) => Object.assign(acc, {[col.columnName]: col.decoder}), {})
-
-  let objects = {_root: {objectId: '_root', type: 'map', props: {}}}
-  let property = null
-
-  while (!col.idActor.done) {
-    const opId = `${col.idCtr.readValue()}@${actorIds[col.idActor.readValue()]}`
-    const action = col.action.readValue(), actionName = ACTIONS[action]
-    if (action % 2 === 0) { // even-numbered actions are object creation
-      const type = OBJECT_TYPE[actionName] || 'unknown'
-      if (type === 'list' || type === 'text') {
-        objects[opId] = {objectId: opId, type, edits: []}
-      } else {
-        objects[opId] = {objectId: opId, type, props: {}}
-      }
-    }
-
-    const objActor = col.objActor.readValue(), objCtr = col.objCtr.readValue()
-    const objId = objActor === null ? '_root' : `${objCtr}@${actorIds[objActor]}`
-    let obj = objects[objId]
-    if (!obj) throw new RangeError(`Operation for nonexistent object: ${objId}`)
-
-    const keyActor = col.keyActor.readValue(), keyCtr = col.keyCtr.readValue()
-    const keyStr = col.keyStr.readValue(), insert = !!col.insert.readValue()
-    const chldActor = col.chldActor.readValue(), chldCtr = col.chldCtr.readValue()
-    const childId = chldActor === null ? null : `${chldCtr}@${actorIds[chldActor]}`
-    const sizeTag = col.valLen.readValue()
-    const rawValue = col.valRaw.readRawBytes(sizeTag >> 4)
-    const value = decodeValue(sizeTag, rawValue)
-    const succNum = col.succNum.readValue()
-    let succ = []
-    for (let i = 0; i < succNum; i++) {
-      succ.push(`${col.succCtr.readValue()}@${actorIds[col.succActor.readValue()]}`)
-    }
-
-    if (!actionName || obj.type === 'unknown') continue
-
-    let key
-    if (obj.type === 'list' || obj.type === 'text') {
-      if (keyCtr === null || (keyCtr === 0 && !insert)) {
-        throw new RangeError(`Operation ${opId} on ${obj.type} object has no key`)
-      }
-      key = insert ? opId : `${keyCtr}@${actorIds[keyActor]}`
-    } else {
-      if (keyStr === null) {
-        throw new RangeError(`Operation ${opId} on ${obj.type} object has no key`)
-      }
-      key = keyStr
-    }
-
-    if (!property || property.objId !== objId || property.key !== key) {
-      let index = 0
-      if (property) {
-        index = property.index
-        if (addPatchProperty(objects, property)) index += 1
-        if (property.objId !== objId) index = 0
-      }
-      property = {objId, key, index, ops: []}
-    }
-    property.ops.push({opId, actionName, value, childId, succ})
-  }
-
-  if (property) addPatchProperty(objects, property)
-  condenseEdits(objects._root)
-  return objects._root
-}
-
 module.exports = {
   COLUMN_TYPE, VALUE_TYPE, ACTIONS, OBJECT_TYPE, DOC_OPS_COLUMNS, CHANGE_COLUMNS, DOCUMENT_COLUMNS,
   encoderByColumnId, decoderByColumnId, makeDecoders, decodeValue,
   splitContainers, encodeChange, decodeChangeColumns, decodeChange, decodeChangeMeta, decodeChanges,
-  encodeDocumentHeader, decodeDocumentHeader, encodeDocument, decodeDocument,
-  getChangeChecksum, appendEdit, constructPatch
+  encodeDocumentHeader, decodeDocumentHeader, decodeDocument
 }
