+++ package.json
@@ -25,7 +25,6 @@
   "license": "MIT",
   "dependencies": {
     "fast-sha256": "^1.3.0",
-    "immutable": "^3.8.2",
     "pako": "^2.0.3",
     "uuid": "^3.4.0"
   },
@@ -35,7 +34,6 @@
     "copyfiles": "^2.4.1",
     "eslint": "^7.24.0",
     "eslint-plugin-compat": "^3.9.0",
-    "jsverify": "^0.8.4",
     "karma": "^6.1.1",
     "karma-chrome-launcher": "^3.1.0",
     "karma-firefox-launcher": "^2.1.0",
